+++ docs/v0.3.0/api-reference/importer-props.mdx
@@ -139,3 +139,14 @@ Use cases for `customKey`:
 - **Session Management**: You can use the `customKey` to implement session-specific storage, allowing you to clear specific import sessions without affecting others.
 
 ---
+
+## translationResources
+
+**Type:** `Record<string, Translation>`
+
+**Required:** No
+
+**Description:**  
+A dictionary of translations for the importer component. The keys are the locale codes, and the values are the translations.
+
+---
