+++ src/i18/types.ts
@@ -1,4 +1,5 @@
 import { ReactNode } from 'preact/compat';
+import type enTranslation from './en/translation.json';
 
 type TranslationValue = string | { [key: string]: TranslationValue };
 export type Translations = Record<string, TranslationValue>;
@@ -15,3 +16,17 @@ export type TranslationContextType = {
   t: (key: string, argumentValues?: ArgumentsTypeText) => string;
   tHtml: (key: string, argumentValues?: ArgumentsTypeHtml) => ReactNode;
 };
+
+type Prettify<T> = {
+  [K in keyof T]: T[K] extends object ? Prettify<T[K]> : T[K];
+} & unknown;
+
+type CreateTranslation<T> = {
+  [K in keyof T]: T[K] extends string
+    ? string
+    : T[K] extends object
+      ? CreateTranslation<T[K]>
+      : T[K];
+};
+
+export type Translation = Prettify<CreateTranslation<typeof enTranslation>>;
